+++ CEX/CJP.cpp
@@ -361,7 +361,7 @@ size_t CJP::ShuffleLoop(uint LowBits, uint MinShift)
 	}
 
 	// add a lower boundary value to ensure we have a minimum RNG loop count
-	return static_cast<size_t>(shuffle + (1 << MinShift));
+	return ((size_t)shuffle + ((size_t)1 << MinShift));
 }
 
 // Shuffle the pool by mixing some value with a bijective function (XOR) into the pool.
