+++ CEX/Keccak512.cpp
@@ -180,7 +180,7 @@ void Keccak512::Reset()
 
 		if (m_parallelProfile.IsParallel())
 		{
-			m_treeParams.NodeOffset() = i;
+			m_treeParams.NodeOffset() = static_cast<uint>(i);
 			Compress(m_treeParams.ToBytes(), 0, m_dgtState[i]);
 		}
 	}
